refactored app to dumb, created content container and content presentational components
